requirements.txt
sentinelsat
shapely
earthpy
descartes
rasterio
pyproj
pandas
geopandas
numpy
IPython
matplotlib
pyperclip
docopt
Rtree
opencv-python
subprocess32
utm
psycopg2
GeoAlchemy2
psycopg-postgis
omegaconf
metaflow
google-cloud-storage
requests
